app: App: run(): prevent panic when player thread returns before main thread
It is possible that the player thread returns before this thread has a
chance to, this would mean that the sender is dropped and thus the
`recv()` would return an `Err`. Since this is the only time an `Err`
would be produced, it is safe to return from this function instead of
hoping it might be able to receive a frame again and thus continuing
within the function.